🚧 refactor using `@nrwl/devkit`
