+++ packages/semver/collection.json
@@ -4,7 +4,7 @@
     "ng-add": {
       "aliases": ["install"],
       "schema": "./src/generators/install/schema.json",
-      "factory": "./src/generators/install/index#install",
+      "factory": "./src/generators/install",
       "description": "Add @jscutlery/semver to the workspace."
     }
   },
@@ -12,7 +12,7 @@
     "ng-add": {
       "aliases": ["install"],
       "schema": "./src/generators/install/schema.json",
-      "factory": "./src/generators/install/index#installSchematic",
+      "factory": "./src/generators/install/compat",
       "description": "Add @jscutlery/semver to the workspace."
     }
   },
