+++ packages/semver/src/executors/version/compat.ts
@@ -1,0 +1,5 @@
+import { convertNxExecutor } from '@nrwl/devkit';
+
+import version from './';
+
+export const versionExecutor = convertNxExecutor(version);
