+++ packages/semver/src/generators/install/compat.ts
@@ -1,0 +1,5 @@
+import { convertNxGenerator } from '@nrwl/devkit';
+
+import install from '.';
+
+export default convertNxGenerator(install);
